Refatorando metodo de Gerar Boleto
Implementando Metodo de gerar Cobranca no Cartao

Implementando Metodo de pagar Cobranca com cartao

Implementando Endpoint para realizar pagamento de cartao

+++ src/main/java/com/pacoprojects/api/integration/juno/cobranca/criar/boleto/RequestCobrancaJunoDto.java
@@ -25,7 +25,7 @@ public record RequestCobrancaJunoDto(
         BigDecimal amount,
 
         @NotNull(message = "Número de parcelas deve ser informado.")
-        Long installMents,
+        Integer installMents,
 
         String reference,
